Rework board convention as always "B to play".
I had previously switched to B/W to play, with the board remaining
identical no matter whose turn it was, because it made SGF parsing
easier. But this defers the board flipping to the NN input stage,
since the NN must always take its input as "B to play". Having two
differing conventions in the code was confusing so now its always
B to play.

(A bug in the previous NN code: all training input was treated as
B to play, even if it was W to play. That meant that the resulting
NN would be very eager to play your best move, if you tenuki'd. I
interpreted this as the NN being very smart/assertive, but it was
just a bug in the training input... q.q)