Clarify semantics of detecting move legality
I had hoped there was an efficient way to generate a list of all
legal moves. Unfortunately, there is not, as far as I can tell.
The issues revolve around efficiently detecting when a move is
suicidal; a move can be suicidal if it's placed directly into an
enemy eye, or if it consumes the last liberty of a group, but only
if that group isn't then merged with another group of that color
having further liberties, and also if no captures happen. The short
story is that you should just play a move, and if that returns None,
your move was illegal.

The ramifications for the MCTS algorithm are that when you expand a
node with all of its children, you can't know efficiently if there
are invalid moves. Additionally, any logic where you select a move
has to loop through all possible moves, checking after each attempted
move whether that move was legal.