Scale down stddev of initial weights according to network width
On average, the sum of N independent variables with stddev S
will be a gaussian variable with stddev S*sqrt(N). With a ReLu
activation function, there is no renormalization of outputs, and
so, the final output activation will be a gaussian variable with
stddev [S * sqrt(N)]^L where L is number of layers.

We have a conv net with 3x3 kernels and depth K, so N in this case
is 9*K. If each of the weights is scaled to be 1/sqrt(9K), then
the output will be again a gaussian variable with stddev S.

Otherwise, we get exponential blowup or shrinkage of the activations
during both forward and backward passes.